requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
aiofiles==23.1.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydantic==1.10.12
openai-whisper
torch
requests
pyannote.audio
librosa
soundfile
numpy
celery[redis]
redis
boto3
